Added commission formula

+++ app/src/main/java/com/mkdev/currencyexchange/extension/Number.kt
@@ -2,7 +2,7 @@ package com.mkdev.currencyexchange.extension
 
 import java.text.DecimalFormat
 
-fun Double.formatTwoDecimalNumber(): Double {
+fun Double.formatTwoDecimalNumber(): Float {
     val decimalFormat = DecimalFormat("#.##")
-    return java.lang.Double.valueOf(decimalFormat.format(this))
+    return java.lang.Float.valueOf(decimalFormat.format(this))
 }
